refactor(l5): move chunk-write, decomposition, secret-patterns, nats-subjects, table-constants from src/ into server/ (issue 864)
L5 of _plans/server-hardening-ladder.md: five leaf modules leave src/ for their
owning server/ directories, with their tests alongside them.

Moved (git mv, contents unchanged):
- src/chunk-write.ts            -> server/capture/chunk-write.ts
- src/chunk-write.pg.test.ts    -> server/capture/chunk-write.pg.test.ts
- src/decomposition.ts          -> server/domain/decomposition.ts
- src/secret-patterns.ts        -> server/security/secret-patterns.ts
- src/nats-subjects.ts          -> server/config/nats-subjects.ts
- src/nats-subjects.test.ts     -> server/config/nats-subjects.test.ts
- src/tools/table-constants.ts  -> server/db/table-constants.ts

server/ importers are repointed at the new paths. Every other importer keeps
its import line untouched: each old src/ path is now a two-line shim that
re-exports the server/ module, so the 26 legacy src/ and scripts/ files never
enter the pre-commit gate. The shims are src/ files and retire with src/ at L6.

Runtime closure of src/ under server/main.ts: 67 -> 62 files, counted with
shims excluded.

Behavior is unchanged; the moved tests pass unmodified.
